refactor: domain 패키지 view 의존성 제거
