+++ src/main/java/lotto/LottoMain.java
@@ -1,40 +1,22 @@
 package lotto;
 
-import lotto.domain.LottoGame;
+import lotto.controller.LottoGameController;
 import lotto.domain.LottoGameResults;
-import lotto.view.InputView;
-import lotto.view.ResultView;
 
-import java.util.Arrays;
-import java.util.Map;
 
 public class LottoMain {
 
     public static void main(String[] args){
 
-        int gameMoney = InputView.getInputMoneyAmount();
+        int gameMoney = LottoGameController.inputGameMoney();
 
-        LottoGame lottoGame = new LottoGame(gameMoney);
-        LottoGameResults lottoGameResults = lottoGame.start();
+        LottoGameResults lottoGameResults = LottoGameController.execute(gameMoney);
 
-        ResultView.getLottoTickets(lottoGameResults);
+        LottoGameController.showLottoTickets(lottoGameResults);
 
+        String lastWinningNumbers = LottoGameController.getLastWinningNumbers();
 
-
-        String lastWinningNumbers = InputView.getInputWinningNumbers();
-
-        lottoGameResults.checkWinningResult(lastWinningNumbers);
-
-        Map<Integer, Integer> resultsMap = lottoGameResults.getWinningResultRecord();
-
-        ResultView.getPrizeResult(resultsMap);
-
-
-
-        double profit = lottoGameResults.getProfit(resultsMap);
-
-        ResultView.getProfitResult(profit);
-
+        LottoGameController.getPrizeResult(lottoGameResults, lastWinningNumbers);
 
     }
 }
