+++ src/test/java/lotto/domain/LottoGameResultsTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.LottoErrorMessage;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
