+++ src/test/java/lotto/domain/PrizeUnitTest.java
@@ -1,7 +1,5 @@
 package lotto.domain;
 
-import lotto.LottoErrorMessage;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
